fashion minist clothes recognition/requirements.txt
tensorflow==2.13.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
streamlit==1.25.0
pillow==10.0.0
scikit-learn==1.3.0
seaborn==0.12.2 